Development

+++ app/Http/Controllers/TaskController.php
@@ -6,24 +6,30 @@ use App\Http\Requests\StoreTaskRequest;
 use App\Models\Task;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TaskController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        $tasks = Task::with('users')
-            ->paginate(1000);
+        $filters = $request->only([
+            'title', 'user_id', 'due_date',
+        ]);
 
-        $tasksByStatus = $tasks->groupBy('status');
 
-        // echo "<pre>";
-        // print_r($tasksByStatus);
-        // exit;
+        $query = Task::with('users')
+            ->filter($filters)
+            ->orderBy('created_at', 'desc')
+            ->get();
 
-        return view('task.index', compact('tasksByStatus', 'tasks'));
+        $tasksByStatus = $query->groupBy('status');
+
+        $users = User::all();
+
+        return view('task.index', compact('tasksByStatus', 'users', 'filters'));
     }
 
     /**
@@ -42,14 +48,15 @@ class TaskController extends Controller
     {
 
         $task = Task::create([
+            'by_user_id' => auth()->id(),
             'title' => $request->title,
             'description' => $request->description,
             'status' => $request->status,
             'due_date' => $request->due_date,
         ]);
 
-        if (isset($request->user_ids)) {
-            $task->users()->attach($request->user_ids, [
+        if (isset($request->user_id)) {
+            $task->users()->attach($request->user_id, [
                 'created_at' => now(),
                 'updated_at' => now(),
             ]);
@@ -71,6 +78,8 @@ class TaskController extends Controller
      */
     public function edit(Task $task)
     {
+        //Gate::authorize("edit", $task);
+
         $users = User::all();
         return view('task.edit', compact('task', 'users'));
     }
@@ -80,6 +89,8 @@ class TaskController extends Controller
      */
     public function update(StoreTaskRequest $request, Task $task)
     {
+        Gate::authorize("update", $task);
+
         // Update the task
         $task->update([
             'title' => $request->title,
@@ -88,7 +99,7 @@ class TaskController extends Controller
             'due_date' => $request->due_date,
         ]);
 
-        $task->users()->sync($request->user_ids);
+        $task->users()->sync($request->user_id);
 
         return redirect()->route('tasks.index')->with('success', 'Task updated successfully!');
     }
@@ -98,6 +109,8 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        Gate::authorize("delete", $task);
+
         $task->delete();
 
         return redirect()->route('tasks.index')->with('success', 'Task deleted successfully!');
